Lots of changes =p
>Protocol - use float(%) to send and receive coordinate msg now
+ Select the server type (JFrame)

+ New user interface for android
+ Connection setup dialog
> extract the android sensor stuff to a class

+++ QuickPointer/src/smallcampus/QuickPointer/net/BaseServer.java
@@ -13,8 +13,8 @@ public abstract class BaseServer {
 		onClientConnected = listener;
 	}
 		
-	protected EventListener<int[]> onCoordinateReceive;
-	public void setOnCoordinateReceiveListener(EventListener<int[]> listener){
+	protected EventListener<float[]> onCoordinateReceive;
+	public void setOnCoordinateReceiveListener(EventListener<float[]> listener){
 		onCoordinateReceive = listener;
 	}
